Further rename Creative Tab API
