+++ fabric-creative-tab-api-v1/src/client/resources/fabric-creative-tab-api-v1.client.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "net.fabricmc.fabric.mixin.creativemodetab.client",
+  "package": "net.fabricmc.fabric.mixin.creativetab.client",
   "compatibilityLevel": "JAVA_25",
   "client": [
     "CreativeModeInventoryScreenMixin"
