+++ fabric-creative-tab-api-v1/src/main/java/net/fabricmc/fabric/api/creativetab/v1/FabricCreativeModeTabOutput.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.creativemodetab.v1;
+package net.fabricmc.fabric.api.creativetab.v1;
 
 import java.util.Arrays;
 import java.util.Collection;
