+++ fabric-creative-tab-api-v1/src/main/java/net/fabricmc/fabric/impl/creativetab/CreativeModeTabEventsImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.creativemodetab;
+package net.fabricmc.fabric.impl.creativetab;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -24,7 +24,7 @@ import org.jspecify.annotations.Nullable;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.CreativeModeTab;
 
-import net.fabricmc.fabric.api.creativemodetab.v1.CreativeModeTabEvents;
+import net.fabricmc.fabric.api.creativetab.v1.CreativeModeTabEvents;
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.EventFactory;
