+++ fabric-creative-tab-api-v1/src/main/java/net/fabricmc/fabric/impl/creativetab/FabricCreativeModeTabBuilderImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.creativemodetab;
+package net.fabricmc.fabric.impl.creativetab;
 
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.CreativeModeTab;
