+++ fabric-creative-tab-api-v1/src/main/java/net/fabricmc/fabric/mixin/creativetab/CreativeModeTabMixin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.creativemodetab;
+package net.fabricmc.fabric.mixin.creativetab;
 
 import java.util.Collection;
 import java.util.LinkedList;
@@ -34,11 +34,11 @@ import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.ItemStack;
 
-import net.fabricmc.fabric.api.creativemodetab.v1.CreativeModeTabEvents;
-import net.fabricmc.fabric.api.creativemodetab.v1.FabricCreativeModeTabOutput;
+import net.fabricmc.fabric.api.creativetab.v1.CreativeModeTabEvents;
+import net.fabricmc.fabric.api.creativetab.v1.FabricCreativeModeTabOutput;
 import net.fabricmc.fabric.api.event.Event;
-import net.fabricmc.fabric.impl.creativemodetab.CreativeModeTabEventsImpl;
-import net.fabricmc.fabric.impl.creativemodetab.FabricCreativeModeTabImpl;
+import net.fabricmc.fabric.impl.creativetab.CreativeModeTabEventsImpl;
+import net.fabricmc.fabric.impl.creativetab.FabricCreativeModeTabImpl;
 
 @Mixin(CreativeModeTab.class)
 abstract class CreativeModeTabMixin implements FabricCreativeModeTabImpl {
