+++ fabric-creative-tab-api-v1/src/main/resources/fabric-creative-tab-api-v1.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "net.fabricmc.fabric.mixin.creativemodetab",
+  "package": "net.fabricmc.fabric.mixin.creativetab",
   "compatibilityLevel": "JAVA_25",
   "mixins": [
     "CreativeModeTabAccessor",
